requirements.txt
sqlmodel~=0.0.23
fastapi~=0.115.11
SQLAlchemy~=2.0.38
pydantic~=2.10.6
pyjwt~=2.10.1